backend/app/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI/ML
torch>=2.2.0
torchvision>=0.17.0
transformers==4.37.0
groq==0.4.1


# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Image Processing
opencv-python==4.9.0.80
Pillow==10.2.0

# Utilities
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0